make tiktoken async (#448)
This PR:
1. Forks tiktoken-node, and uses our old tiktoken-rs fork.
2. Makes priompt async, which should reduce the choppiness in our server
streams.

Had to do the first one because the public tiktoken-node handles
tokenization incorrectly wrt special tokens, and doesn't allow you to
specify im_start etc as special tokens. Decided to do the second one
because it was convenient to do it all together and it was something
that had been nagging me a lot.