[specedits] simple proof-of-concept (#463)
Problems encountered:

1. GPT-4 often likes to change inconsequential things like whitespace at
the top of the file. If this is consistent it would slow down the edits.
2. It is remarkably hard to make GPT-4 only make edits in a particular
small region of a file, without also doing inconsequential things like
removing commented-out code and cleaning it up.
3. When streaming, OpenAI buffers the tokens to make sure it always
gives you back correct utf-8. When inspecting tokens individually, that
isn't possible, and so we have to have special logic for dealing with
the case when one utf-8 code point is represented by two or more tokens.